Use color instead of opacity to show link strength
This requires us to first construct two sets of link objects (lines and joints)
that we want to render, then sort each set by the strength of the associated link
and render the joints on a separate layer above the lines – to ensure that lighter
lines don't overdraw darker ones and that we can always use joint colors to assess
pairwise link strength even when overdrawing is going on. As a bonus, we also get
to reliably draw the moves themselves as *another* layer on top of all the links,
so links don't ever overdraw moves anymore.